+++ bean-searcher/src/main/java/com/ejlchina/searcher/implement/StrNumFieldConvertor.java
@@ -15,21 +15,26 @@ import java.math.BigInteger;
 public class StrNumFieldConvertor implements FieldConvertor.BFieldConvertor {
 
     @Override
-    public boolean supports(FieldMeta meta, Class<?> valueType, Class<?> targetType) {
-        return valueType == String.class && (
-               targetType == int.class || targetType == Integer.class ||
-               targetType == long.class || targetType == Long.class ||
-               targetType == float.class || targetType == Float.class ||
-               targetType == double.class || targetType == Double.class ||
-               targetType == short.class || targetType == Short.class ||
-               targetType == byte.class || targetType == Byte.class ||
-               targetType == BigDecimal.class || targetType == BigInteger.class
-        );
+    public boolean supports(FieldMeta meta, Class<?> valueType) {
+        if (valueType == String.class) {
+            Class<?> targetType = meta.getType();
+            return (
+                    targetType == int.class || targetType == Integer.class ||
+                    targetType == long.class || targetType == Long.class ||
+                    targetType == float.class || targetType == Float.class ||
+                    targetType == double.class || targetType == Double.class ||
+                    targetType == short.class || targetType == Short.class ||
+                    targetType == byte.class || targetType == Byte.class ||
+                    targetType == BigDecimal.class || targetType == BigInteger.class
+            );
+        }
+        return false;
     }
 
     @Override
-    public Object convert(FieldMeta meta, Object value, Class<?> targetType) {
+    public Object convert(FieldMeta meta, Object value) {
         String number = (String) value;
+        Class<?> targetType = meta.getType();
         if (targetType == int.class || targetType == Integer.class) {
             return Integer.parseInt(number);
         }